refactor: move loaders out of backends
